The message context menu is now added to only the front message subcell. I also added an open thread action to it. The input delegation has been refactored on the message cell to accommodate both taps and long presses.
